moving generator to backend index.js
